updated with registration logic db calls and jstl tag libraries introduced for dynamic population of drop down values (sample is done with states drop down in registration page).
Always call servlets and they generate data and forward request to jsp
so as the page dynamic data is generated on the fly.

stil to consider:
cache the master data and update it periodically. 
send mail on registration.
show the unique user error with proper message (now it shows db error
directly).

clean up for the jquery resources. added the zip fils into resources
folder. need to test and remote not required files from scripts folder